*)fixed bug in PerformanceMemory_p.java which caused negative memory-values on big peers see http://www.yacy-forum.de/viewtopic.php?t=2370
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@2091 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ htroot/PerformanceMemory_p.java
@@ -63,9 +63,9 @@ public class PerformanceMemory_p {
     private static final int MB = 1024 * KB;
     private static Map defaultSettings = null;
         
-    private static int[]    slt,chk;
+    private static long[]    slt,chk;
     private static String[] ost;
-    private static int      req, usd, bst, god;
+    private static long     req, usd, bst, god;
     
     private static long usedTotal, currTotal, dfltTotal, goodTotal, bestTotal;
